+++ src/components/schedule-tables/index.ts
@@ -1,2 +1,1 @@
-export * from "./ScheduleTable";
 export * from "./ScheduleTables";
